feat!: support dev register plugins
