+++ Cargo.toml
@@ -47,6 +47,7 @@ tempfile = { version = "3.15.0", default-features = false }
 tokio = { version = "1.45.1", default-features = false, features = ["full"] }
 tracing = { version = "0.1.41", default-features = false, features = ["attributes"] }
 tracing-subscriber = { version = "0.3.19", default-features = false, features = ["env-filter", "ansi", "time", "json"] }
+uuid = { version = "1.17.0", default-features = false, features = ["std"] }
 wasm-pkg-client = { version = "0.10.0", default-features = false }
 wasm-pkg-core = { version = "0.10.0", default-features = false }
 wasmcloud-runtime = { git = "https://github.com/brooksmtownsend/wasmCloud", version = "0.11.0", branch = "feat/minimal-runtime-handler" }
@@ -58,6 +59,7 @@ wit-component = { version = "0.235.0", default-features = false }
 wit-parser = { version = "0.235.0", default-features = false }
 
 [dev-dependencies]
+bytes = { version = "1", default-features = false }
 http = { version = "1.3.1", default-features = false }
 http-body-util = { version = "0.1.3", default-features = false }
 tempfile = { version = "3.0", default-features = false }
