+++ crates/wash/src/runtime/mod.rs
@@ -1,8 +1,7 @@
-use anyhow::{self, Context as _, ensure};
+use anyhow::{self, Context as _};
 use sha2::{Digest as _, Sha256};
 use std::collections::HashMap;
 use std::fmt::Debug;
-use std::path::PathBuf;
 use std::sync::Arc;
 use std::thread::JoinHandle;
 use tokio::sync::RwLock;
@@ -13,44 +12,36 @@ use wasmcloud_runtime::capability::logging::logging::Level;
 use wasmcloud_runtime::component::BaseCtx;
 use wasmtime::{
     AsContextMut as _, StoreContextMut,
-    component::{
-        Component, Instance, InstancePre, Linker, ResourceAny, ResourceTable, ResourceType, Val,
-        types::ComponentItem,
-    },
+    component::{Component, Instance, InstancePre, Linker, ResourceTable, types::ComponentItem},
 };
 use wasmtime_wasi::{IoImpl, IoView, WasiCtx, WasiCtxBuilder, WasiView};
 use wasmtime_wasi_http::{WasiHttpCtx, WasiHttpView};
 
 use crate::runtime::{
     types::{WasmcloudWashCtx, WasmcloudWashImpl, WasmcloudWashView},
-    value::{is_host_resource_type, lift, lower},
+    wasm::link_imports_plugin_exports,
 };
 
 pub mod bindings;
 pub(crate) mod types;
-mod value;
-
-#[derive(Hash, Eq, PartialEq, Debug, Clone)]
-pub enum CapabilityPlugin {
-    Blobstore,
-}
+mod wasm;
 
 pub struct Ctx {
-    pub wash_ctx: WasmcloudWashCtx,
+    /// Unique identifier for this component context. Primarily used for logging, debugging, and
+    /// plugin instance management.
+    pub id: String,
     pub table: wasmtime::component::ResourceTable,
     pub ctx: WasiCtx,
     pub http: WasiHttpCtx,
+    pub wash: WasmcloudWashCtx,
 
     /// `wasi:config/runtime` uses this hashmap to retrieve runtime configuration
     pub runtime_config: Arc<RwLock<HashMap<String, String>>>,
-    /// `wasi:blobstore/blobstore` uses this path to store blobs
-    pub blobstore_root: Option<PathBuf>,
 }
 
 pub struct CtxBuilder {
     ctx: WasiCtx,
     runtime_config: Option<HashMap<String, String>>,
-    blobstore_root: Option<PathBuf>,
 }
 
 impl CtxBuilder {
@@ -61,7 +52,6 @@ impl CtxBuilder {
                 .inherit_stderr()
                 .build(),
             runtime_config: None,
-            blobstore_root: None,
         }
     }
 
@@ -75,15 +65,9 @@ impl CtxBuilder {
         self
     }
 
-    pub fn with_blobstore_root(mut self, path: Option<PathBuf>) -> Self {
-        self.blobstore_root = path;
-        self
-    }
-
     pub fn build(self) -> Ctx {
         Ctx {
             runtime_config: Arc::new(RwLock::new(self.runtime_config.unwrap_or_default())),
-            blobstore_root: self.blobstore_root,
             ctx: self.ctx,
             ..Default::default()
         }
@@ -105,14 +89,14 @@ impl Ctx {
 impl Default for Ctx {
     fn default() -> Self {
         Self {
-            wash_ctx: WasmcloudWashCtx::default(),
+            id: uuid::Uuid::now_v7().to_string(),
+            wash: WasmcloudWashCtx::default(),
             table: ResourceTable::new(),
             ctx: WasiCtxBuilder::new()
                 .args(&["main.wasm"])
                 .inherit_stderr()
                 .build(),
             http: WasiHttpCtx::new(),
-            blobstore_root: None,
             runtime_config: Arc::new(RwLock::new(HashMap::new())),
         }
     }
@@ -121,7 +105,7 @@ impl Default for Ctx {
 impl Debug for Ctx {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.debug_struct("Ctx")
-            .field("wash_ctx", &self.wash_ctx)
+            .field("wash_ctx", &self.wash)
             .field("table", &self.table)
             .finish()
     }
@@ -144,7 +128,7 @@ impl WasiHttpView for Ctx {
 }
 impl WasmcloudWashView for Ctx {
     fn ctx(&mut self) -> &mut WasmcloudWashCtx {
-        &mut self.wash_ctx
+        &mut self.wash
     }
 }
 
@@ -218,217 +202,6 @@ pub async fn prepare_component_plugin(
     Ok(component)
 }
 
-/// This function plugs a components imports with the exports of other components
-/// that are already loaded in the plugin system.
-pub fn link_imports_plugin_exports(
-    linker: &mut Linker<Ctx>,
-    component: &Component,
-    plugin_manager: Arc<DevPluginManager>,
-) -> anyhow::Result<()> {
-    let ty = component.component_type();
-    let imports: Vec<_> = ty.imports(component.engine()).collect();
-    for (import_name, import_item) in imports.into_iter() {
-        match import_item {
-            ComponentItem::ComponentInstance(import_instance_ty) => {
-                debug!(name = import_name, "processing component instance import");
-                let (plugin_component, instance_idx) = {
-                    let Some(plugin_component) = plugin_manager.get_component(import_name) else {
-                        debug!(name = import_name, "import not found in plugins, skipping");
-                        continue;
-                    };
-                    let Some((ComponentItem::ComponentInstance(_), idx)) =
-                        plugin_component.export_index(None, import_name)
-                    else {
-                        debug!(name = import_name, "skipping non-instance import");
-                        continue;
-                    };
-                    (plugin_component, idx)
-                };
-                debug!(name = import_name, index = ?instance_idx, "found import at index");
-
-                // Preinstantiate the instance so we can use it later
-                let pre = plugin_manager.preinstantiate_instance(linker, import_name)?;
-
-                let mut linker_instance = match linker.instance(import_name) {
-                    Ok(i) => i,
-                    Err(e) => {
-                        debug!(name = import_name, error = %e, "error finding instance in linker, skipping");
-                        continue;
-                    }
-                };
-
-                for (export_name, export_ty) in
-                    import_instance_ty.exports(plugin_component.engine())
-                {
-                    match export_ty {
-                        ComponentItem::ComponentFunc(_func_ty) => {
-                            let (item, func_idx) = match plugin_component
-                                .export_index(Some(&instance_idx), export_name)
-                            {
-                                Some(res) => res,
-                                None => {
-                                    debug!(
-                                        name = import_name,
-                                        fn_name = export_name,
-                                        "failed to get export index, skipping"
-                                    );
-                                    continue;
-                                }
-                            };
-                            ensure!(
-                                matches!(item, ComponentItem::ComponentFunc(..)),
-                                "expected function export, found other"
-                            );
-                            debug!(
-                                name = import_name,
-                                fn_name = export_name,
-                                "linking function import"
-                            );
-                            let import_name: Arc<str> = import_name.into();
-                            let export_name: Arc<str> = export_name.into();
-                            let pre = pre.clone();
-                            let plugin_manager = plugin_manager.clone();
-                            linker_instance
-                                .func_new_async(
-                                    &export_name.clone(),
-                                    move |mut store, params, results| {
-                                        let import_name = import_name.clone();
-                                        let export_name = export_name.clone();
-                                        let plugin_manager = plugin_manager.clone();
-                                        let pre = pre.clone();
-                                        Box::new(async move {
-                                            let instance = plugin_manager
-                                                .get_or_instantiate_instance(
-                                                    &mut store,
-                                                    pre,
-                                                    &import_name,
-                                                )
-                                                .await?;
-
-                                            let func = instance
-                                                .get_func(&mut store, func_idx)
-                                                .context("function not found")?;
-                                            trace!(
-                                                name = %import_name,
-                                                fn_name = %export_name,
-                                                ?params,
-                                                "lowering params"
-                                            );
-                                            let mut params_buf = Vec::with_capacity(params.len());
-                                            for param_val in params.iter() {
-                                                params_buf.push(
-                                                    lower(&mut store, param_val)
-                                                        .context("failed to lower parameter")?,
-                                                );
-                                            }
-                                            trace!(
-                                                name = %import_name,
-                                                fn_name = %export_name,
-                                                ?params_buf,
-                                                "invoking dynamic export"
-                                            );
-
-                                            let mut results_buf =
-                                                vec![Val::Bool(false); results.len()];
-                                            func.call_async(
-                                                &mut store,
-                                                &params_buf,
-                                                &mut results_buf,
-                                            )
-                                            .await
-                                            .context("failed to call function")?;
-                                            trace!(
-                                                name = %import_name,
-                                                fn_name = %export_name,
-                                                ?results_buf,
-                                                "lifting results"
-                                            );
-                                            for (i, result_val) in
-                                                results_buf.into_iter().enumerate()
-                                            {
-                                                results[i] = lift(&mut store, result_val)
-                                                    .context("failed to lift result")?;
-                                            }
-                                            trace!(
-                                                name = %import_name,
-                                                fn_name = %export_name,
-                                                ?results,
-                                                "invoked dynamic export"
-                                            );
-
-                                            func.post_return_async(&mut store)
-                                                .await
-                                                .context("failed to execute post-return")?;
-                                            Ok(())
-                                        })
-                                    },
-                                )
-                                .expect("failed to create async func");
-                        }
-                        ComponentItem::Resource(resource_ty) => {
-                            if is_host_resource_type(resource_ty) {
-                                debug!(name = import_name, resource = ?resource_ty, "skipping host resource type");
-                                continue;
-                            }
-
-                            let (item, _idx) = match plugin_component
-                                .export_index(Some(&instance_idx), export_name)
-                            {
-                                Some(res) => res,
-                                None => {
-                                    debug!(
-                                        name = import_name,
-                                        resource = export_name,
-                                        "failed to get resource index, skipping"
-                                    );
-                                    continue;
-                                }
-                            };
-                            let ComponentItem::Resource(_) = item else {
-                                debug!(
-                                    name = import_name,
-                                    resource = export_name,
-                                    "expected resource export, found non-resource, skipping"
-                                );
-                                continue;
-                            };
-
-                            debug!(name = import_name, resource = export_name, ty = ?resource_ty, "linking resource import");
-                            linker_instance
-                                .resource(export_name, ResourceType::host::<ResourceAny>(), |_, _| Ok(()))
-                                .with_context(|| {
-                                    format!(
-                                        "failed to define resource import: {import_name}.{export_name}"
-                                    )
-                                })
-                                .unwrap_or_else(|e| {
-                                    debug!(name = import_name, resource = export_name, error = %e, "error defining resource import, skipping");
-                                });
-                        }
-                        _ => {
-                            trace!(
-                                name = import_name,
-                                fn_name = export_name,
-                                "skipping non-function non-resource import"
-                            );
-                            continue;
-                        }
-                    }
-                }
-            }
-            ComponentItem::Resource(resource_ty) => {
-                debug!(
-                    name = import_name,
-                    ty = ?resource_ty,
-                    "component import is a resource, which is not supported in this context. skipping."
-                );
-            }
-            _ => continue,
-        }
-    }
-    Ok(())
-}
-
 /// Creates a WebAssembly component from bytes and compiles using [Runtime].
 ///
 /// This function is used to prepare a component for use in the plugin system, adding in the following interfaces:
@@ -459,12 +232,11 @@ pub async fn prepare_component_dev(
                 .context("failed to link `wasi:http`")?;
             wasmcloud_runtime::capability::logging::logging::add_to_linker(linker, |ctx| ctx)
                 .context("failed to link `wasi:logging/logging`")?;
-            add_to_linker_async(linker)
-                .context("failed to link `wasmcloud:wash` host interface")?;
-
-            // TODO: Use a component that sources env vars for this
             wasmcloud_runtime::capability::config::runtime::add_to_linker(linker, |ctx| ctx)
                 .context("failed to link `wasi:config/runtime`")?;
+
+            add_to_linker_async(linker)
+                .context("failed to link `wasmcloud:wash` host interface")?;
 
             // Link additional configured plugins
             link_imports_plugin_exports(linker, component, plugin_manager)
@@ -495,11 +267,6 @@ where
     val
 }
 
-// what we basically want is a lookup from instance to Component to start, but then
-// to manage a single Instance for that component. E.g. blobstore,container,types should all go to the same instance once instantiated
-
-// I think we can have the plugins managed by a struct, and use methods for properly looking these up
-
 // A hash of the component bytes
 type ComponentKey = String;
 
@@ -510,9 +277,10 @@ pub struct DevPluginManager {
     interface_map: HashMap<String, ComponentKey>,
     // component key → compiled component
     components: HashMap<ComponentKey, Arc<Component>>,
-
-    // component key → instantiated instance (created lazily)
-    instances: RwLock<HashMap<ComponentKey, Arc<Instance>>>,
+    // Instances is keyed differently than components, as it's keyed by a specific Instance of a component
+    // This allows us to instantiate the same component multiple times with different configurations.
+    // component id → component key -> instantiated instance (created lazily)
+    instances: RwLock<HashMap<String, HashMap<ComponentKey, Arc<Instance>>>>,
 }
 
 impl DevPluginManager {
@@ -548,7 +316,7 @@ impl DevPluginManager {
         linker: &Linker<Ctx>,
         name: &str,
     ) -> anyhow::Result<Arc<InstancePre<Ctx>>> {
-        info!(name, "preinstantiating instance for component");
+        trace!(name, "preinstantiating instance for component");
         let component = self.get_component(name).context("component not found")?;
         let instance = linker.instantiate_pre(&component)?;
         Ok(Arc::new(instance))
@@ -563,20 +331,41 @@ impl DevPluginManager {
         let key = self
             .interface_map
             .get(name)
-            .ok_or_else(|| anyhow::anyhow!("no component implements interface {name}"))?;
-        if let Some(instance) = self.instances.read().await.get(key) {
-            return Ok(instance.clone());
+            .context("component not found")?;
+
+        let id = store.data().id.clone();
+
+        if let Some(instances) = self.instances.read().await.get(&id) {
+            if let Some(instance) = instances.get(key) {
+                trace!(id, "found existing instance for component");
+                return Ok(instance.clone());
+            }
         }
 
-        let instance = pre.instantiate_async(store.as_context_mut()).await?;
-        let instance = Arc::new(instance);
-
+        let instance = Arc::new(pre.instantiate_async(store.as_context_mut()).await?);
+        trace!(id, key, "instantiated new instance for component");
         self.instances
             .write()
             .await
-            .insert(key.to_string(), instance.clone());
+            .entry(id.to_string())
+            .and_modify(|e| {
+                e.insert(key.clone(), instance.clone());
+            })
+            .or_insert(HashMap::from([(key.clone(), instance.clone())]));
 
         Ok(instance)
+    }
+
+    /// Clone this DevManager and clear the instances map, removing all instantiated plugin components.
+    ///
+    /// Must be done before each instantiation of a component to ensure that the
+    /// plugins are using the same store as the instantiated component.
+    pub fn clear_instances(self: Arc<Self>) -> Arc<Self> {
+        Arc::new(Self {
+            interface_map: self.interface_map.clone(),
+            components: self.components.clone(),
+            instances: RwLock::new(HashMap::new()),
+        })
     }
 }
 
@@ -611,7 +400,7 @@ mod test {
     use tempfile::TempDir;
     use wasmtime_wasi::{DirPerms, FilePerms};
     use wasmtime_wasi_http::{
-        bindings::http::types::Scheme,
+        bindings::http::types::{ErrorCode, Scheme},
         body::{HostIncomingBody, HyperIncomingBody},
         types::HostIncomingRequest,
     };
@@ -648,6 +437,9 @@ mod test {
 
     #[tokio::test]
     async fn can_instantiate_http_component() -> anyhow::Result<()> {
+        let _ = tracing_subscriber::fmt()
+            .with_env_filter(tracing_subscriber::EnvFilter::from_default_env())
+            .try_init();
         let wasm = std::fs::read("./tests/fixtures/http_hello_world_rust.wasm")?;
 
         let (runtime, _handle) = new_runtime().await?;
@@ -706,14 +498,14 @@ mod test {
 
     #[tokio::test]
     async fn can_instantiate_blobstore_component() -> anyhow::Result<()> {
-        tracing_subscriber::fmt()
+        let _ = tracing_subscriber::fmt()
             .with_env_filter(tracing_subscriber::EnvFilter::from_default_env())
-            .init();
+            .try_init();
 
         let (runtime, _handle) = new_runtime().await?;
 
         // This component has a `wasi:blobstore/blobstore` IMPORT
-        let wasm = std::fs::read("./tests/fixtures/blobstore_from_fs.wasm")?;
+        let wasm = std::fs::read("./tests/fixtures/http_blobstore.wasm")?;
         let blobstore_plugin = std::fs::read("./tests/fixtures/blobstore_filesystem.wasm")?;
 
         let mut plugin_manager = DevPluginManager::default();
@@ -730,7 +522,7 @@ mod test {
         let base_ctx = Ctx::builder()
             .with_wasi_ctx(
                 WasiCtxBuilder::new()
-                    .preopened_dir(tmp_path, "/tmp", DirPerms::all(), FilePerms::all())
+                    .preopened_dir(&tmp_path, "/tmp", DirPerms::all(), FilePerms::all())
                     .context("failed to create preopened dir")?
                     .build(),
             )
@@ -747,10 +539,16 @@ mod test {
             .context("failed to pre-instantiate `wasi:http/incoming-handler`")?;
 
         let data = store.data_mut();
+        let body = HyperIncomingBody::new(
+            // One gigabyte of data. Larger than this will hang on the get_data TODO: investigate
+            http_body_util::Full::new(hyper::body::Bytes::from("0123456789".repeat(100_000)))
+                .map_err(|e| ErrorCode::InternalError(Some(e.to_string())))
+                .boxed(),
+        );
         let request: ::http::Request<wasmtime_wasi_http::body::HyperIncomingBody> =
             http::Request::builder()
                 .uri("http://localhost:8080")
-                .body(HyperIncomingBody::default())
+                .body(body)
                 .context("failed to create request")?;
 
         let (parts, body) = request.into_parts();
@@ -780,8 +578,9 @@ mod test {
             .await
             .context("failed to collect body bytes")?
             .to_bytes();
-        let body_str = String::from_utf8_lossy(&body);
-        assert_eq!(body_str, "my-container-real");
+        assert_eq!(body.len(), 1_000_000);
+
+        eprintln!("roundtrip streamed data: {}", body.len());
 
         Ok(())
     }
